+++ test/htmlMinify.js
@@ -12,6 +12,8 @@ describe('HTML Minification Tests', function () {
   const appDir = path.join(__dirname, 'app/htmlMinifier')
   const appConfig = {
     appDir,
+    expressSession: false,
+    secretsDir: 'secrets',
     port: 41002,
     logging: {
       methods: {