+++ test/htmlValidator.js
@@ -37,6 +37,8 @@ describe('validator usage', () => {
     // spin up the roosevelt app
     roosevelt({
       mode: 'development',
+      expressSession: false,
+      secretsDir: 'secrets',
       makeBuildArtifacts: false,
       port: 40001,
       logging: {