+++ test/httpsOptions.js
@@ -46,7 +46,6 @@ describe('HTTPS Server Options Tests', async () => {
     const fs = require('fs')
     const key = pem.private
     const cert = pem.cert
-    const keyFolder = './test/app/certs'
     const path = require('path')
 
     if (!appDir) {
@@ -66,25 +65,27 @@ describe('HTTPS Server Options Tests', async () => {
     const p12Cert = forge.asn1.toDer(p12Info).getBytes()
 
     try {
-      if (!fs.existsSync(keyFolder)) {
-        fs.mkdirSync(keyFolder)
+      const secretDir = './test/app/' + config.secretsDir
+
+      if (!fs.existsSync(secretDir)) {
+        fs.mkdirSync(secretDir)
       }
 
-      fs.writeFileSync('./test/app/certs/key.pem', key, err => {
+      fs.writeFileSync(secretDir + '/key.pem', key, err => {
         if (err) {
           console.error(err)
         }
         // file written successfully
       })
 
-      fs.writeFileSync('./test/app/certs/cert.pem', cert, err => {
+      fs.writeFileSync(secretDir + '/cert.pem', cert, err => {
         if (err) {
           console.error(err)
         }
         // file written successfully
       })
 
-      fs.writeFileSync('./test/app/certs/cert.p12', p12Cert, err => {
+      fs.writeFileSync(secretDir + '/cert.p12', p12Cert, err => {
         if (err) {
           console.error(err)
         }
@@ -108,7 +109,7 @@ describe('HTTPS Server Options Tests', async () => {
 
   // after all tests clean up the test app directory
   after(function (done) {
-    fs.rmSync('./test/app/certs', { recursive: true, force: true })
+    fs.rmSync('./test/app/' + config.secretsDir, { recursive: true, force: true })
 
     cleanupTestApp(appDir, (err) => {
       if (err) {
@@ -120,7 +121,7 @@ describe('HTTPS Server Options Tests', async () => {
   })
 
   it('should create a https server when https.enable is set to true', function () {
-    app({ appDir, ...config })
+    app({ appDir: './test/app', ...config })
 
     // test assertion
     assert(stubHttpsServer.called, 'https.Server was not called')
@@ -129,8 +130,7 @@ describe('HTTPS Server Options Tests', async () => {
   it('should create a http server and a https server when the https.force param is false', function () {
     // change config
     config.https.force = false
-
-    app({ appDir, ...config })
+    app({ appDir: './test/app', ...config })
 
     // test assertion
     assert(stubHttpServer.called, 'http.Server was not called')
@@ -141,8 +141,7 @@ describe('HTTPS Server Options Tests', async () => {
     // change config
     config.https.force = true
     config.https.enable = false
-
-    app({ appDir, ...config })
+    app({ appDir: './test/app', ...config })
 
     // test assertion
     assert(stubHttpServer.called, 'http.Server was not called')
@@ -154,26 +153,26 @@ describe('HTTPS Server Options Tests', async () => {
   it('should not create a http.Server when the https.force param is set to true', function () {
     // change config
     config.https.force = true
-
-    app({ appDir, ...config })
+    app({ appDir: './test/app', ...config })
 
     // test assertion
     assert(stubHttpServer.notCalled, 'http.Server called despite force flag')
   })
 
   it('should start a https server using the given p12 file and passphrase if the p12Path param is set to a file path string and the passphrase is set', function () {
-    app({ appDir, ...config })
+    app({ appDir: './test/app', ...config })
 
     // test assertions
     assert(typeof stubHttpsServer.args[0][0].cert === 'undefined', 'https.Server had cert when using p12')
     assert(typeof stubHttpsServer.args[0][0].key === 'undefined', 'https.Server had key when using p12')
-    assert(fs.existsSync(path.join(appDir, './../../../test/app/certs/cert.p12')) === true, 'file at config p12 file path does not exist')
+    assert(fs.existsSync('./test/app/secrets/cert.p12') === true, 'file at config p12 file path does not exist')
     assert(stubHttpsServer.args[0][0].passphrase === config.https.passphrase, 'https.Server passphrase did not match config passphrase')
     assert(stubHttpsServer.called, 'https.Server was not called')
   })
 
   it('should start a https server using the given p12 buffer and passphrase if the p12.p12Path param is set to a PKCS#12 formatted buffer and passphrase is set', function () {
-    const p12text = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.p12'), 'utf8')
+    const appDir = './test/app'
+    const p12text = fs.readFileSync('./test/app/secrets/cert.p12', 'utf8')
     config.https.authInfoPath.p12.p12Path = p12text
     app({ appDir, ...config })
 
@@ -186,8 +185,9 @@ describe('HTTPS Server Options Tests', async () => {
   })
 
   it('should start a https server using the given certAndKey.cert and certAndKey.key params if they are set with file path strings', function () {
-    const keytext = fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))
-    const certext = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))
+    const appDir = './test/app'
+    const keytext = fs.readFileSync('./test/app/secrets/key.pem')
+    const certext = fs.readFileSync('./test/app/secrets/cert.pem')
     // change config - unset p12Path
     config.https.authInfoPath.p12.p12Path = ''
     config.https.authInfoPath.authCertAndKey.key = keytext
@@ -203,17 +203,18 @@ describe('HTTPS Server Options Tests', async () => {
   })
 
   it('should start a https server using the given certAndKey.cert and certAndKey.key params if one is a certificate string and one is a file path', function () {
-    const keytext = fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))
-    const certext = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))
+    const appDir = './test/app'
+    const keytext = fs.readFileSync('./test/app/secrets/key.pem')
+    const certext = fs.readFileSync('./test/app/secrets/cert.pem')
 
     // change config - change key to a string
     config.https.authInfoPath.p12.p12Path = null
-    config.https.authInfoPath.authCertAndKey.key = fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))
-    config.https.authInfoPath.authCertAndKey.cert = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))
+    config.https.authInfoPath.authCertAndKey.key = fs.readFileSync('./test/app/secrets/key.pem')
+    config.https.authInfoPath.authCertAndKey.cert = fs.readFileSync('./test/app/secrets/cert.pem')
 
     app({ appDir, ...config })
 
-    // // test assertions
+    // test assertions
     assert(stubHttpsServer.args[0][0].key.equals(keytext), 'https.Server key file did not match supplied')
     assert(stubHttpsServer.args[0][0].cert.equals(certext), 'https.Server cert file did not match supplied')
     assert(typeof stubHttpsServer.args[0][0].pfx === 'undefined', 'https.Server options had pfx set when p12Path was null')
@@ -221,12 +222,13 @@ describe('HTTPS Server Options Tests', async () => {
   })
 
   it('should start a https server using the given certAndKey.cert and certAndKey.key params if they are set with certificate strings', function () {
-    const keytext = fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))
-    const certext = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))
+    const appDir = './test/app'
+    const keytext = fs.readFileSync('./test/app/secrets/key.pem')
+    const certext = fs.readFileSync('./test/app/secrets/cert.pem')
 
     // change config - change cert to a  string
-    config.https.authInfoPath.authCertAndKey.key = fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))
-    config.https.authInfoPath.authCertAndKey.cert = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))
+    config.https.authInfoPath.authCertAndKey.key = fs.readFileSync('./test/app/secrets/key.pem')
+    config.https.authInfoPath.authCertAndKey.cert = fs.readFileSync('./test/app/secrets/cert.pem')
 
     app({ appDir, ...config })
 
@@ -239,29 +241,29 @@ describe('HTTPS Server Options Tests', async () => {
 
   it('should start a https server using the certificate authority certificate from a file if the caCert param is set with a file path', function () {
     config.https.authInfoPath.p12.p12Path = null
-
+    const appDir = './test/app'
     app({ appDir, ...config })
 
     // test assertion
-    assert(config.https.caCert === 'test/app/certs/cert.pem', 'https.Server CA did not match supplied CA')
+    assert(config.https.caCert === 'cert.pem', 'https.Server CA did not match supplied CA')
     assert(stubHttpsServer.called, 'https.Server was not called')
   })
 
   it('should start a https using a certificate chain as an array of certificates when the caCert param is set with an array of file paths', function () {
     // change config
-    config.https.caCert = [path.join(appDir, './../../../test/app/certs/cert.pem'), path.join(appDir, './../../../test/app/certs/key.pem')]
+    config.https.caCert = [path.join(appDir, './../../../test/app/secrets/cert.pem'), path.join(appDir, './../../../test/app/secrets/key.pem')]
 
     app({ appDir, ...config })
 
     // test assertions
-    assert(stubHttpsServer.args[0][0].ca[0].equals(fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))), 'https.Server CA (1) did not match supplied CA')
-    assert(stubHttpsServer.args[0][0].ca[1].equals(fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))), 'https.Server CA (1) did not match supplied CA')
+    assert(stubHttpsServer.args[0][0].ca[0].equals(fs.readFileSync(path.join(appDir, './../../../test/app/secrets/cert.pem'))), 'https.Server CA (1) did not match supplied CA')
+    assert(stubHttpsServer.args[0][0].ca[1].equals(fs.readFileSync(path.join(appDir, './../../../test/app/secrets/key.pem'))), 'https.Server CA (1) did not match supplied CA')
     assert(stubHttpsServer.called, 'https.Server was not called')
   })
 
   it('should start a https server using the caCert param directly if it is set as a certificate string', function () {
     // change config
-    config.https.caCert = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'), 'UTF8')
+    config.https.caCert = fs.readFileSync(path.join(appDir, './../../../test/app/secrets/cert.pem'), 'UTF8')
 
     app({ appDir, ...config })
 
@@ -271,8 +273,8 @@ describe('HTTPS Server Options Tests', async () => {
   })
 
   it('should start a https server using the caCert certificate chain if the caCert param is set as an array of certificate strings', function () {
-    const ca1 = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))
-    const ca2 = fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))
+    const ca1 = fs.readFileSync(path.join(appDir, './../../../test/app/secrets/cert.pem'))
+    const ca2 = fs.readFileSync(path.join(appDir, './../../../test/app/secrets/key.pem'))
 
     // change config
     config.https.caCert = [ca1, ca2]
@@ -286,16 +288,16 @@ describe('HTTPS Server Options Tests', async () => {
   })
 
   it('should start a https server using the caCert certificate chain if it the caCert param set as an array of mixed file paths and certificate strings', function () {
-    const ca1 = fs.readFileSync(path.join(appDir, './../../../test/app/certs/cert.pem'))
+    const ca1 = fs.readFileSync(path.join(appDir, './../../../test/app/secrets/cert.pem'))
 
     // change config
-    config.https.caCert = [ca1, path.join(appDir, './../../../test/app/certs/key.pem')]
+    config.https.caCert = [ca1, path.join(appDir, './../../../test/app/secrets/key.pem')]
 
     app({ appDir, ...config })
 
     // test assertions
     assert(stubHttpsServer.args[0][0].ca[0].equals(ca1), 'https.Server CA (1) did not match supplied CA')
-    assert(stubHttpsServer.args[0][0].ca[1].equals(fs.readFileSync(path.join(appDir, './../../../test/app/certs/key.pem'))), 'https.Server CA (2) did not match supplied CA')
+    assert(stubHttpsServer.args[0][0].ca[1].equals(fs.readFileSync(path.join(appDir, './../../../test/app/secrets/key.pem'))), 'https.Server CA (2) did not match supplied CA')
     assert(stubHttpsServer.called, 'https.Server was not called')
   })